services/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolade Fajimi",email = "fajimikolade23@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "grpcio (>=1.78.0,<2.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "cachetools (>=7.0.3,<8.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis (>=7.3.0,<8.0.0)",
    "protobuf (>=7.34.0,<8.0.0)",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

# ← no [[tool.poetry.source]] block at all